requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
PyJWT==2.11.0
razorpay==2.0.0
reportlab==4.4.9
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
python-dotenv==1.2.1
